kinexis_backend/requirements.txt
# Core dependencies for Kinexis Backend
Flask==2.3.2
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0

# MediaPipe for pose detection
mediapipe==0.10.7
opencv-python==4.8.1.78
numpy==1.24.3

# Database
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.19

# Video processing
Pillow==10.0.0

# Environment variables
python-dotenv==1.0.0

# PDF Report Generation
reportlab==4.0.4
matplotlib==3.7.2

# Data processing
pandas==2.0.3

# Testing (optional)
pytest==7.4.0
pytest-flask==1.2.0